pyproject.toml
[tool.poetry]
name = "python-fun"
version = "0.1.0"
description = ""
authors = ["Ariel Parra <darkaico@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
jupyterlab = "^3.3.3"
matplotlib = "^3.5.1"
tweepy = "^4.8.0"
python-dotenv = "^0.20.0"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
black = {version = "^22.3.0", allow-prereleases = true}
ipdb = "^0.13.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 99
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
